# enroute-cp/docs/swagger.yaml
definitions:
...
  webhttp.Secret:
    properties:
      secret_cert:
        type: string
      secret_key:
        type: string
      secret_name:
        type: string
      secret_sni:
        type: string
    type: object
...
info:
  contact:
    name: API Support
    url: https://saaras.io/
  description: APIs to configure multiple envoy proxies
  license: {}
  title: Enroute API
  version: "1.0"
paths:
...
  /secret:
    get:
      consumes:
      - application/json
      description: Get a list of all secrets for all services
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: ""
      security:
      - ApiKeyAuth: []
      summary: List all secrets
      tags:
      - secret
    post:
      consumes:
      - application/json
      parameters:
      - description: Secret to create
        in: body
        name: Secret
        required: true
        schema:
          $ref: '#/definitions/webhttp.Secret'
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: ""
      security:
      - ApiKeyAuth: []
      summary: Create a secret
      tags:
      - secret
  /secret/{secret_name}:
    delete:
      consumes:
      - application/json
      parameters:
      - description: Name of secret
        in: path
        name: secret_name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: ""
      security:
      - ApiKeyAuth: []
      summary: Delete a secret
      tags:
      - secret
  /secret/{secret_name}/cert:
    post:
      consumes:
      - application/json
      description: |-
        Set the secret cert from file
        Example curl -X POST -F 'Secret_cert=@certificate.pem' http://localhost:1323/secret/testsecret/cert | python -m json.tool
      parameters:
      - description: Name of secret
        in: path
        name: secret_name
        required: true
        type: string
      - description: Location of file holding the secret cert
        in: formData
        name: secret_cert
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: ""
      security:
      - ApiKeyAuth: []
      summary: Set the secret cert from file
      tags:
      - secret
  /secret/{secret_name}/key:
    post:
      consumes:
      - application/json
      description: |-
        Set the secret key from file
        Example curl -X POST -F 'Secret_key=@private_key.pem' http://localhost:1323/secret/testsecret/key | python -m json.tool
      parameters:
      - description: Name of secret
        in: path
        name: secret_name
        required: true
        type: string
      - description: Location of file holding the secret key
        in: formData
        name: secret_key
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: ""
      security:
      - ApiKeyAuth: []
      summary: Set the secret key from file
      tags:
      - secret
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"